[BUGFIX] Cleanup types

+++ dll_injection/src/map_buffer_into_process.h
@@ -23,7 +23,7 @@ PVOID map_buffer_into_process1(HANDLE hProcess, LPBYTE buffer, SIZE_T buffer_siz
     printf("Section handle: %p\n", hSection);
 
     PVOID sectionBaseAddress = NULL;
-    ULONG viewSize = 0;
+    SIZE_T viewSize = 0;
     SECTION_INHERIT inheritDisposition = ViewShare; //VIEW_SHARE
 
     // map the section in context of current process:
